have token stuff as an interface going to settings
