+++ src/OneOff/ToolkitTokenParameters.cs
@@ -1,0 +1,25 @@
+using System.Security.Cryptography.X509Certificates;
+using FatCat.Toolkit.Console;
+using FatCat.Toolkit.Web.Api;
+using Humanizer;
+using Microsoft.IdentityModel.Tokens;
+
+namespace OneOff;
+
+public class ToolkitTokenParameters : IToolkitTokenParameters
+{
+	public TokenValidationParameters Get()
+	{
+		ConsoleLog.WriteCyan("Getting token parameters");
+
+		var cert = new X509Certificate2(@"C:\DevelopmentCert\DevelopmentCert.pfx", "basarab_cert");
+
+		return new TokenValidationParameters
+				{
+					IssuerSigningKey = new X509SecurityKey(cert),
+					ValidAudience = "https://foghaze.com/Brume",
+					ValidIssuer = "FogHaze",
+					ClockSkew = 10.Seconds()
+				};
+	}
+}
