+++ src/ToolKit/Web/Api/ToolkitTokenParameterGenerator.cs
@@ -1,29 +1,11 @@
-﻿using System.Security.Cryptography.X509Certificates;
-using FatCat.Toolkit.Console;
-using Humanizer;
-using Microsoft.IdentityModel.Tokens;
+﻿using Microsoft.IdentityModel.Tokens;
 
 namespace FatCat.Toolkit.Web.Api;
 
+/// <summary>
+/// Used to get TokenParameters for the Api
+/// </summary>
 public interface IToolkitTokenParameters
 {
 	TokenValidationParameters Get();
-}
-
-public class ToolkitTokenParameters : IToolkitTokenParameters
-{
-	public TokenValidationParameters Get()
-	{
-		ConsoleLog.WriteCyan("Getting token parameters");
-
-		var cert = new X509Certificate2(@"C:\DevelopmentCert\DevelopmentCert.pfx", "basarab_cert");
-
-		return new TokenValidationParameters
-				{
-					IssuerSigningKey = new X509SecurityKey(cert),
-					ValidAudience = "https://foghaze.com/Brume",
-					ValidIssuer = "FogHaze",
-					ClockSkew = 10.Seconds()
-				};
-	}
 }
